Added -ld flag, that will consider compared value as the same value if the difference between them is less than 1E (ld value)

+++ src/fstcomp/statfldx.F90
@@ -32,7 +32,7 @@
       type(fst_record ) :: record  
 
       integer i,j,k
-      real(kind=real32) :: sum,moy,var,rmin,rmax
+      real(kind=real64) :: sum,moy,var,rmin,rmax
       integer imin,jmin,kmin,imax,jmax,kmax,kind,dat2,dat3
       character(len=15) :: Level
       REAL      rlevel
